lib: convert prefixlen to 16-bit integer

+++ zebra/connected.c
@@ -278,7 +278,7 @@ void connected_up(struct interface *ifp, struct connected *ifc)
 
 /* Add connected IPv4 route to the interface. */
 void connected_add_ipv4(struct interface *ifp, int flags, struct in_addr *addr,
-			uint8_t prefixlen, struct in_addr *broad,
+			uint16_t prefixlen, struct in_addr *broad,
 			const char *label, uint32_t metric)
 {
 	struct prefix_ipv4 *p;
@@ -477,7 +477,7 @@ static void connected_delete_helper(struct connected *ifc, struct prefix *p)
 
 /* Delete connected IPv4 route to the interface. */
 void connected_delete_ipv4(struct interface *ifp, int flags,
-			   struct in_addr *addr, uint8_t prefixlen,
+			   struct in_addr *addr, uint16_t prefixlen,
 			   struct in_addr *broad)
 {
 	struct prefix p, d;
@@ -503,7 +503,7 @@ void connected_delete_ipv4(struct interface *ifp, int flags,
 
 /* Add connected IPv6 route to the interface. */
 void connected_add_ipv6(struct interface *ifp, int flags, struct in6_addr *addr,
-			struct in6_addr *broad, uint8_t prefixlen,
+			struct in6_addr *broad, uint16_t prefixlen,
 			const char *label, uint32_t metric)
 {
 	struct prefix_ipv6 *p;
@@ -561,7 +561,7 @@ void connected_add_ipv6(struct interface *ifp, int flags, struct in6_addr *addr,
 }
 
 void connected_delete_ipv6(struct interface *ifp, struct in6_addr *address,
-			   struct in6_addr *broad, uint8_t prefixlen)
+			   struct in6_addr *broad, uint16_t prefixlen)
 {
 	struct prefix p, d;
 	struct connected *ifc;
